+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/TestData.kt
@@ -36,6 +36,7 @@ internal const val INVALID_FILE_URL_PATH = "https://78d85043-d04f-486a-8212-bf26
 
 internal const val TEST_CUSTOM_PERMISSION_RESULT = "{\"rakuten.miniapp.user.USER_NAME\":\"DENIED\"}"
 internal const val TEST_USER_NAME = "test_user_name"
+internal const val TEST_PROFILE_PHOTO = "data:image/png;base64,encodedValue"
 
 internal val TEST_MA = MiniAppInfo(
     id = TEST_MA_ID,