+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### 2.x.x (in progress)
 **SDK**
-- **Feature:** Added `getUserName()` interface in JavaScript bridge for receiving user name if the permission has been granted.
+- **Feature:** Added `getUserName()` interface in JavaScript bridge for receiving user name if the required permission has been granted e.g. "rakuten.miniapp.user.USER_NAME".
+- **Feature:** Added `getProfilePhoto()` interface in JavaScript bridge for receiving profile photo url if the required permission has been granted e.g. "rakuten.miniapp.user.PROFILE_PHOTO".
 
 ### 2.2.0 (2020-10-02)
 **SDK**